Many warning fixes for clang 3.8. More work on checking bounds when computing elastic material properties from the a mechanical behaviour. Check errno value at the end of the computation of material properties. Change member names related to bounds in the MaterialPropertyDescription structure

+++ mfront/include/MFront/BoundsDescription.hxx
@@ -75,7 +75,7 @@ namespace mfront{
     BoundsType     boundsType  = LowerAndUpper;
     long double    lowerBound  = std::numeric_limits<long double>::min();
     long double    upperBound  = std::numeric_limits<long double>::max();
-    unsigned int   lineNumber  = 0u;
+    size_t         lineNumber  = 0u;
 
   protected:
